refactor(LottoTest): 중복생성된 테스트 통합
